app/requirements.txt
django==5.0.1
python-dotenv==1.0.0
mysqlclient==2.2.1
django-mptt==0.15.0
django-seed==0.3.1
Faker==22.0.0